requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-cors-headers==2.1.0
django-extensions==1.9.8
django-rest-swagger==2.1.2
djangorestframework==3.7.3
djangorestframework-simplejwt==3.2.3
fuzzywuzzy==0.17.0
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
numpy==1.14.5
openapi-codec==1.3.2
pandas==0.23.3
psycopg2==2.7.3.2
pydotplus==2.0.2
PyJWT==1.5.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.3
requests==2.18.4
simplejson==3.12.0
six==1.11.0
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==1.0.5
